refactor: introduce heartlytics package
