+++ heartlytics/__init__.py
@@ -1,0 +1,13 @@
+"""Heartlytics application package."""
+
+from .app import app, create_app
+from .extensions import db, login_manager
+from . import models
+
+__all__ = [
+    "app",
+    "create_app",
+    "db",
+    "login_manager",
+    "models",
+]
