+++ heartlytics/blueprints/routes/debug.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from flask import Blueprint, render_template, request, redirect, url_for, flash, current_app
 from flask_login import login_required
-from auth.decorators import require_roles
-from services.email import get_events
+from ...auth.decorators import require_roles
+from ...services.email import get_events
 
 
 debug_bp = Blueprint("debug", __name__, url_prefix="/debug")
