+++ heartlytics/models.py
@@ -1,0 +1,25 @@
+"""Convenience imports for ORM models."""
+
+from .app import (
+    User,
+    AuditLog,
+    PasswordResetRequest,
+    MFAEmailChallenge,
+    Role,
+    UserRole,
+    Patient,
+    Prediction,
+    ClusterSummary,
+)
+
+__all__ = [
+    "User",
+    "AuditLog",
+    "PasswordResetRequest",
+    "MFAEmailChallenge",
+    "Role",
+    "UserRole",
+    "Patient",
+    "Prediction",
+    "ClusterSummary",
+]
